% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analyze_resource_utilization}
\alias{analyze_resource_utilization}
\title{Analyze Resource Utilization}
\usage{
analyze_resource_utilization(employee_data, project_data)
}
\arguments{
\item{employee_data}{DataFrame containing employee data}

\item{project_data}{DataFrame containing project data}
}
\value{
List containing utilization analysis
}
\description{
Analyze current resource allocation and utilization
}
