% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_business_days}
\alias{calculate_business_days}
\title{Calculate Business Days}
\usage{
calculate_business_days(start_date, end_date)
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}
}
\value{
Number of business days
}
\description{
Calculate business days between two dates
}
\examples{
start_date <- as.Date("2023-01-02")
end_date <- as.Date("2023-01-06")
calculate_business_days(start_date, end_date)
}
