% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_percentile}
\alias{calculate_percentile}
\title{Calculate Percentile}
\usage{
calculate_percentile(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector of percentile ranks (0-100)
}
\description{
Calculate percentile rank within a group
}
\examples{
values <- c(10, 20, 30, 40, 50)
calculate_percentile(values)
}
