% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_currency}
\alias{format_currency}
\title{Format Currency}
\usage{
format_currency(x, currency = "EUR ")
}
\arguments{
\item{x}{Numeric vector}

\item{currency}{Currency symbol}
}
\value{
Character vector with formatted currency
}
\description{
Format numbers as currency
}
\examples{
format_currency(1000)
format_currency(c(1000, 2500), "EUR ")
}
