% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_dashboard.R
\name{get_dashboard_url}
\alias{get_dashboard_url}
\title{Get Dashboard URL}
\usage{
get_dashboard_url(port = 3838, host = "127.0.0.1")
}
\arguments{
\item{port}{Port number}

\item{host}{Host address}
}
\value{
Character string with dashboard URL
}
\description{
Get Dashboard URL
}
