% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalTechAnalytics-package.R, R/analytics.R
\docType{package}
\name{GlobalTechAnalytics-package}
\alias{GlobalTechAnalytics-package}
\alias{analytics}
\title{GlobalTechAnalytics: Resource Analytics for GlobalTech Enterprises}
\description{
Resource analytics platform with predictive models and interactive dashboards.

Core analytics functions for workforce insights
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Anna-KG/GlobalTechAnalytics}
  \item Report bugs at \url{https://github.com/Anna-KG/GlobalTechAnalytics/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/Anna-KG/GlobalTechAnalytics}
  \item Report bugs at \url{https://github.com/Anna-KG/GlobalTechAnalytics/issues}
}

}
\author{
\strong{Maintainer}: Anna Kędzierska \email{a.kedzierska@student.uw.edu.pl}

}
\keyword{internal}
