% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_logger}
\alias{setup_logger}
\title{Setup Logger}
\usage{
setup_logger(name = "GlobalTechAnalytics")
}
\arguments{
\item{name}{Logger name}
}
\value{
List with logging functions
}
\description{
Creates a simple logging function
}
\examples{
logger <- setup_logger("MyApp")
logger$info("This is an info message")
}
